Fix hash table chaining
Current hash table handling breaks third (Brent's) invariant for
hash table collision chaining. When a node was chosen to be a
chain master (master = hash(key)) there is no way to known whether
this node already in some chain (of different master) or not because
the key, which is contained in the node, could be already collected by GC.
So any attempt to form a relinking algorithm out is worthless by
design - we could not have any clue what is really stored in the
keys (both successors and predecessors).
In order to fix this issue I suggest to introduce a prev link which
turns a collision chain in a double linked list. In any case
(node is alive, node is dead, the key still persists or is collected)
the prev link gives us a strong knowledge whether a node is a master
node or linked in a chain. This knowledge allows us to follow the
Brent invariant in any circumstances and provide some lookup
optimization, so if the first (key master) node has a predecessor
(prev is not NULL) then the chain (and table) does not contain given key.
Also the patch got rid of linear scanning in lj_tab_newkey as
predecessor is already known using the prev value.

In order to make code insert and relink simpler a new key is going to
be put into a key master node even when this node is not empty and
key:value pair from which are going to be shifted into the allocated
node (freenode).

Relates to: LuaJIT gh-494